python script
The aim is to use machine learning techniques to develop an algorithm that is able to
predict the life expectancy of a device. This will be achieved by setting up an evaluation framework
and using exploratory data analysis (EDA) to select adequate performance measures and randomly
split the data into training and test sets. Then use regularisation and hyperparameter tuning to
improve the performance and analyse the results.